Add homes from Account page
Created a popup modal for the user to add homes to within the account page. Creates JS functions for the AddHome modal in styles.js
Modal --per requirements of HTML -- is located in the footer.inc.php source code before the <script> tags.
Adding a home through the modal adds the home through to the database and shows up in the account page.